type: recalibrate integer types to match lotusd/bitcoinsuite canonical sources
Every domain field now uses the Rust integer type that matches its canonical
lotusd FlatBuffers schema, rather than widening everything to i64. Type
conversions are confined to two infrastructure boundaries: SQLite reads
(i64 → target type with debug_assert guard) and JSON API serialization
(u64 → String for safe roundtrips).

Type map:
  uint64 (lotusd) → u64 (domain): template_id, template_epoch, coinbase_value
  int32  (lotusd) → i32 (domain): height (block height, chain tip)
  int64  (lotusd) → i64 (domain): CAmount/monetary fields (already correct)

Changes by file:

  docs/adrs/006-canonical-type-alignment.md
    - New ADR documenting the decision, rationale, considered options, and
      consequences

  src/accounting/share_repository.rs
    - Share.template_id: i64 → u64
    - Share.template_epoch: i64 → u64
    - build_dedupe_key() params match new types
    - create_test_share() helper updated

  src/accounting/round_repository.rs
    - Round.start_template_id: i64 → u64
    - Round.end_template_id: Option<i64> → Option<u64>
    - All function signatures and SQLite reads updated with debug_assert guards

  src/accounting/found_block_repository.rs
    - FoundBlock.height: i64 → i32
    - FoundBlock.template_id: Option<i64> → Option<u64>
    - FoundBlock.coinbase_value: i64 → u64
    - All SQLite reads (get_by_hash, get_by_round_id, list) use explicit
      i64 → target type casts with debug_assert guards

  src/accounting/accounting_event_repository.rs
    - AccountingEvent.template_id: Option<i64> → Option<u64>
    - AccountingEvent.template_epoch: Option<i64> → Option<u64>
    - AccountingEvent.height: Option<i64> → Option<i32>
    - SQLite reads updated with debug_assert guards

  src/accounting/service.rs
    - record_share(), resolve_round_for_template(), close_round(),
      record_found_block(), check_maturation(), reconcile_found_blocks(),
      get_or_create_current_round() — all parameter types updated
    - Single documented boundary conversion: coinbase_value (u64) →
      gross_reward (i64) at the NNG→CAmount boundary

  src/accounting/chain_tip.rs
    - Internal AtomicU64 → AtomicI32 to match lotusd's int32_t height
    - Methods: new(), update_block_connected(), update_block_disconnected(),
      get() — all i32

  src/payout/plan.rs
    - PayoutPlan.block_height: i64 → i32
    - build_payout_plan() parameter updated

  src/stratum_protocol/job.rs
    - parse_template_metadata_from_job_id() returns (u64, u64)

  src/stratum_protocol/server.rs
    - Removed all job.template_id as i64, job.template_epoch as i64,
      job.height as i64, job.coinbase_value as i64 casts

  src/node_integration/json_rpc/client.rs
    - getblockhash() height param: i64 → i32

  src/node_integration/nng/consumer.rs
    - Removed event.height as u64 cast (ChainTip now accepts i32)
    - Removed chain_tip.get() as i64 cast

  src/main.rs
    - Removed tip_height as i64 casts at both call sites

  src/http_api/routes/rounds.rs, blocks.rs, shares.rs
    - API DTOs serialize u64 fields as JSON String (safe for values > 2^53)
    - API DTOs serialize i32 fields as JSON number (exact representation)

+++ src/accounting/chain_tip.rs
@@ -1,4 +1,4 @@
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::{AtomicI32, Ordering};
 use std::sync::Arc;
 
 /// Tracks the latest known chain tip height from NNG pub/sub events.
@@ -16,19 +16,19 @@ use std::sync::Arc;
 /// Initial value of 0 is harmless — before the first template refresh
 /// no blocks exist, so no immature blocks are queried.
 #[derive(Clone, Debug)]
-pub struct ChainTip(Arc<AtomicU64>);
+pub struct ChainTip(Arc<AtomicI32>);
 
 impl ChainTip {
     /// Create a new chain tip tracker starting at `initial`.
-    pub fn new(initial: u64) -> Self {
-        Self(Arc::new(AtomicU64::new(initial)))
+    pub fn new(initial: i32) -> Self {
+        Self(Arc::new(AtomicI32::new(initial)))
     }
 
     /// Update on a `BlockConnected` event.
     ///
     /// The tip can only advance — if `height` is lower than the current
     /// value (e.g. stale event), it is ignored.
-    pub fn update_block_connected(&self, height: u64) {
+    pub fn update_block_connected(&self, height: i32) {
         let prev = self.0.load(Ordering::Relaxed);
         if height > prev {
             self.0.store(height, Ordering::Relaxed);
@@ -41,7 +41,7 @@ impl ChainTip {
     /// current tip. Disconnections of older blocks (e.g. deep reorg) are
     /// ignored — those are handled by the orphaning logic in
     /// `handle_block_disconnected`.
-    pub fn update_block_disconnected(&self, height: u64) {
+    pub fn update_block_disconnected(&self, height: i32) {
         let prev = self.0.load(Ordering::Relaxed);
         if height == prev && prev > 0 {
             self.0.store(prev - 1, Ordering::Relaxed);
@@ -49,7 +49,7 @@ impl ChainTip {
     }
 
     /// Get the current chain tip height.
-    pub fn get(&self) -> u64 {
+    pub fn get(&self) -> i32 {
         self.0.load(Ordering::Relaxed)
     }
 }
